requirements.txt
pip~=25.0
filelock~=3.16.1
requests~=2.32.3
colorama~=0.4.6
PyYAML~=6.0.2
numpy~=2.2.1
charset-normalizer~=3.4.0
regex~=2024.11.6
typing_extensions~=4.12.2
tqdm~=4.67.1
packaging~=24.2
urllib3~=2.3.0
fsspec~=2024.12.0
certifi~=2024.12.14
idna~=3.10
safetensors~=0.4.5
transformers~=4.47.1
huggingface-hub~=0.27.0
tokenizers~=0.21.0
confluent-kafka~=2.7.0
python-dotenv~=1.0.0

